// src/Services/const/enum.ts
export enum Topic {
  SOCIETY = "Xã hội",
  SPORT = "Thể thao",
  ENTERTAINMENT = "Giải trí",
  EDUCATION = "Giáo dục",
  SCIENCE = "Khoa học",
  TECHNOLOGY = "Công nghệ",
  ECONOMY = "Kinh tế",
  POLITICS = "Chính trị",
  WORLD = "Thế giới",
  LAW = "Pháp luật",
  HEALTH = "Sức khỏe",
  CULTURE = "Văn hóa",
  LIFE = "Đời sống",
  TRAVEL = "Du lịch",
  FASHION = "Thời trang",
  FOOD = "Ẩm thực",
  CAR = "Xe",
  MOVIE = "Phim",
  MUSIC = "Nhạc",
  BOOK = "Sách",
}